+++ examples/hello_world.rs
@@ -1,7 +1,7 @@
 use janetrs::client::{Error, JanetClient};
 
 fn main() -> Result<(), Error> {
-    let client = JanetClient::init()?.with_default_env();
+    let client = JanetClient::init_with_default_env()?;
 
     client.run("(print `Hello from Janet!`)")?;